[cleanup] Re-organized c++ unittest code
