+++ unittest/test_actions.cpp
@@ -10,7 +10,7 @@
 #define BOOST_TEST_NO_MAIN
 #define BOOST_TEST_ALTERNATIVE_INIT_API
 
-#include "action_factory.hpp"
+#include "factory/action.hpp"
 #include "unittest_common.hpp"
 
 using namespace boost::unit_test;
